V1.0
Currently supports a default mode and a custom mode. User can optionally input the length, width, and difficulty of the game. User can check tiles, place flags, and win/lose the game. A true game board is generated with mine quantities, blank spaces, and mines placed in a 2d character array. An active game board keeps track of what the user sees and interacts with and starts off with each tile hidden by default.

View code for more details on the implementation. Abstraction and generalization optimizations are to come. Contract information as well.